+++ src/include/migraphx/pad_calc.hpp
@@ -5,14 +5,12 @@
 
 namespace migraphx {
 inline namespace MIGRAPHX_INLINE_NS {
-namespace op {
 
 inline size_t calculate_padding(size_t weight_dim, size_t dilation)
 {
     return (dilation * (weight_dim - 1)) / 2;
 }
 
-} // namespace op
 } // namespace MIGRAPHX_INLINE_NS
 } // namespace migraphx
