+++ src/tf/tf.cpp
@@ -17,7 +17,7 @@
 #include <migraphx/instruction.hpp>
 #include <migraphx/config.hpp>
 #include <migraphx/tf.hpp>
-#include <migraphx/op/pad_calc.hpp>
+#include <migraphx/pad_calc.hpp>
 
 namespace migraphx {
 inline namespace MIGRAPHX_INLINE_NS {
@@ -326,8 +326,8 @@ struct tf_parser
             if(pad_mode.find("SAME") != std::string::npos)
             {
                 op.padding_mode = op::padding_mode_t::same;
-                op.padding[0]   = op::calculate_padding(weight_h, op.dilation[0]);
-                op.padding[1]   = op::calculate_padding(weight_w, op.dilation[1]);
+                op.padding[0]   = calculate_padding(weight_h, op.dilation[0]);
+                op.padding[1]   = calculate_padding(weight_w, op.dilation[1]);
             }
             else if(pad_mode.find("VALID") != std::string::npos)
             {
@@ -409,8 +409,8 @@ struct tf_parser
             if(pad_mode.find("SAME") != std::string::npos)
             {
                 op.padding_mode = op::padding_mode_t::same;
-                op.padding[0]   = op::calculate_padding(weight_h, op.dilation[0]);
-                op.padding[1]   = op::calculate_padding(weight_w, op.dilation[1]);
+                op.padding[0]   = calculate_padding(weight_h, op.dilation[0]);
+                op.padding[1]   = calculate_padding(weight_w, op.dilation[1]);
             }
             else if(pad_mode.find("VALID") != std::string::npos)
             {
@@ -573,8 +573,8 @@ struct tf_parser
             if(pad_mode.find("SAME") != std::string::npos)
             {
                 op.padding_mode = op::padding_mode_t::same;
-                op.padding[0]   = op::calculate_padding(op.lengths[0], 1);
-                op.padding[1]   = op::calculate_padding(op.lengths[1], 1);
+                op.padding[0]   = calculate_padding(op.lengths[0], 1);
+                op.padding[1]   = calculate_padding(op.lengths[1], 1);
             }
             else if(pad_mode.find("VALID") != std::string::npos)
             {
